% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assqlddl.R
\name{as.sql_ddl}
\alias{as.sql_ddl}
\title{Help page title}
\usage{
as.sql_ddl(col_spec)
}
\arguments{
\item{col_spec}{Contains the name of the SQL Data type
and special formatting instructions in paranthesis}
}
\value{
Return a strftime formatted string that
can be used by parse_date()
}
\description{
A couple of lines of description about the function(s).
If you want to include code, use \code{my_code()}.
}
\note{
Describe how the algorithm works, or if the function has
any quirks here.
}
\examples{
#R code run by the example function
\dontrun{
#R code that isn't run by example or when the package is built
}
}
\author{
Stefano Da Ros
}
\keyword{misc}
